Rename original->raw, documentation->reference
There was a syntax conflict between the old `original()` build directive and the new `original` build directive suboption.
That conflict can't easily be reconciled using logic because one is a string and the other is a boolean. So I've changed the name to raw.

Documentation was a little too opinionated, since we might reference files that are not documentation. Switched to the more general word "reference"